+++ projects/ngx-tour-console/src/lib/tour-anchor.directive.ts
@@ -1,5 +1,5 @@
-import { TourAnchorDirective, TourService, IStepOption } from 'ngx-tour-core';
-import { Directive, Input, OnDestroy, OnInit } from '@angular/core';
+import {TourAnchorDirective, TourService, IStepOption, isInViewport, ElementSides} from 'ngx-tour-core';
+import {Directive, ElementRef, Input, OnDestroy, OnInit} from '@angular/core';
 
 @Directive({
   selector: '[tourAnchor]',
@@ -7,7 +7,10 @@ import { Directive, Input, OnDestroy, OnInit } from '@angular/core';
 export class TourAnchorConsoleDirective implements OnInit, OnDestroy, TourAnchorDirective {
   @Input() public tourAnchor: string;
 
-  constructor(private tourService: TourService) {}
+  constructor(
+      private tourService: TourService,
+      private element: ElementRef
+  ) {}
 
   public ngOnInit(): void {
     this.tourService.register(this.tourAnchor, this);
@@ -18,6 +21,16 @@ export class TourAnchorConsoleDirective implements OnInit, OnDestroy, TourAnchor
   }
 
   public showTourStep(step: IStepOption): void {
+    const htmlElement: HTMLElement = this.element.nativeElement;
+
+    if (!step.preventScrolling) {
+      if (!isInViewport(htmlElement, ElementSides.Bottom)) {
+        htmlElement.scrollIntoView(false);
+      } else if (!isInViewport(htmlElement, ElementSides.Top)) {
+        htmlElement.scrollIntoView(true);
+      }
+    }
+
     console.group(step.title);
     console.log(step.content);
     console.log(`${step.placement || 'above'} ${this.tourAnchor}`);