+++ projects/ngx-tour-core/jest.config.js
@@ -1,0 +1,11 @@
+module.exports = {
+    globals: {
+        'ts-jest': {
+            diagnostics: false
+        }
+    },
+    transform: {
+        '^.+\\.ts$': 'ts-jest'
+    },
+    verbose: true
+};