+++ projects/ngx-tour-ngx-popper/CHANGELOG.md
@@ -1,3 +1,12 @@
+<a name="4.2.0"></a>
+
+# [4.2.0](https://github.com/isaacplmann/ngx-tour) (2020-08-17)
+
+### Features
+
+- Replaced `withinviewport()` function from 3rd party library with our own custom `isInViewport()` function to fix 
+CommonJS warning in Angular 10, reduce bundle size and to remove non-peer dependency
+
 <a name="4.1.1"></a>
 
 # [4.1.1](https://github.com/isaacplmann/ngx-tour) (2019-12-07)